Break out element display and click handlers into hooks
